refactor(util): fine tune ImeUtils (InputMethodUtils)
